+++ CHANGELOG.md
@@ -1,7 +1,15 @@
+## 1.1.2
+
+* Get country by `dialCode`.
+
+## 1.1.1
+
+* Add method get all countries.
+
 ## 1.1.0
 
-* Add flutter dependency
-* Add country flags
+* Add flutter dependency.
+* Add country flags.
 
 ## 1.0.0